backend/app/requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
transformers
sentence-transformers
weaviate-client
python-dotenv
numpy

